+++ aluguel-ms/src/main/java/com/example/aluguel_ms/ciclista/repository/CiclistaRepository.java
@@ -4,6 +4,6 @@ import com.example.aluguel_ms.ciclista.model.Ciclista;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface CiclistaRepository extends JpaRepository<Ciclista, Long> {
+public interface CiclistaRepository extends JpaRepository<Ciclista, Integer> {
     Optional<Ciclista> findByEmail(String email);
 }
